requirements.txt
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
coverage==5.5
Flask==2.0.0
gunicorn==20.1.0
idna==2.10
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-cov==2.12.1
python-dotenv==0.17.1
requests==2.25.1
toml==0.10.2
urllib3==1.26.5
Werkzeug==2.0.1
